requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
python-dateutil==2.8.1
    # via -r requirements.in
six==1.15.0
    # via python-dateutil
